Create followers.py
Searches the list of followers and returns follows to anyone the bot is not already following back. Sends a thank you message to each new follow, then notifies the monitoring account of the total number of followers added.
Uncomment the last code block to unfollow anyone who is not following the bot.